First functional version of qraise with --infrastructure. Toy example of infrastructure on examples/infrastructure

+++ src/cli/qraise.cpp
@@ -15,6 +15,7 @@
 #include "qraise/simple_conf_qraise.hpp"
 #include "qraise/cc_conf_qraise.hpp"
 #include "qraise/qc_conf_qraise.hpp"
+#include "qraise/infrastructure_conf_qraise.hpp"
 
 #include "logger.hpp"
 
@@ -23,6 +24,7 @@ namespace {
 }
 
 using namespace std::literals;
+using namespace cunqa;
 
 namespace {
 
@@ -172,26 +174,32 @@ int main(int argc, char* argv[])
     const char* store = std::getenv("STORE");
     std::string info_path = std::string(store) + "/.cunqa/qpus.json";
 
-    // Setting and checking mode and family name, respectively
-    std::string mode = args.cloud ? "cloud" : "hpc";
-    std::string family = args.family_name;
-    if (exists_family_name(family, info_path)) { //Check if there exists other QPUs with same family name
-        LOGGER_ERROR("There are QPUs with the same family name as the provided: {}.", family);
-        std::system("rm qraise_sbatch_tmp.sbatch");
-        return -1;
+    if (args.infrastructure.has_value()) {
+            std::ofstream sbatchFile("qraise_sbatch_tmp.sbatch");
+            write_sbatch_infrastructure_file(sbatchFile, args);
+    } else {
+        // Setting and checking mode and family name, respectively
+        std::string mode = args.cloud ? "cloud" : "hpc";
+        std::string family = args.family_name;
+        if (exists_family_name(family, info_path)) { //Check if there exists other QPUs with same family name
+            LOGGER_ERROR("There are QPUs with the same family name as the provided: {}.", family);
+            std::system("rm qraise_sbatch_tmp.sbatch");
+            return -1;
+        }
+
+        // Writing the sbatch file
+        std::ofstream sbatchFile("qraise_sbatch_tmp.sbatch");
+        write_sbatch_header(sbatchFile, args);
+        write_env_variables(sbatchFile);
+        write_run_command(sbatchFile, args, mode);
+        sbatchFile.close();
+
     }
 
-    // Writing the sbatch file
-    std::ofstream sbatchFile("qraise_sbatch_tmp.sbatch");
-    write_sbatch_header(sbatchFile, args);
-    write_env_variables(sbatchFile);
-    write_run_command(sbatchFile, args, mode);
-    sbatchFile.close();
-
-    
     // Executing and deleting the file
     std::system("sbatch qraise_sbatch_tmp.sbatch");
     std::system("rm qraise_sbatch_tmp.sbatch");
-
+    
+    
     return 0;
 }
